feat: adds view storage agreement on my offers also moves agreements utils to more generic location
